Keep stable order of tasks on the taskbar.
The tasks are sorted in order of their start up time on the taskbar even though
new apps are started and other apps are terminated. This is a must for easy
navigation and UI best practise – icons are not continuously moving without
user's intention.

The order of the tasks is defined base on stable sequence IDs of their windows
(see https://developer.gnome.org/meta/stable/MetaWindow.html#meta-window-get-stable-sequence).